automate management of elasticsearch plugin repository
This uses maven itself to download the required jars from Maven central,
verify their GPG signatures against a list of trusted keys and upload them
to our archiva repository.

The elasticsearch plugins themselves are not hosted on Maven central and
their Maven metadata is not available. They are downloaded from
https://artifacts.elastic.co instead and no GPG check is performed.

The deployment repo is updated with the version of the plugins that is
configured in the pom.xml. The changes are not committed automatically.

Change-Id: I4951607ee83901496baadbfcfb69f5be5f3248c8