Use a custom repository-s3 snapshot
Includes a one-line change from the upstream sources to invoke
AmazonS3ClientBuilder.enablePathStyleAccess when building the
s3 client. This is necessary for compatability with the wmf
swift cluster which has an s3 compatability layer, but does
not appear to support virtual host-style access outside
k8s.

This hackery is not necessary in 7.10, elasticsearch removed
the ability to set this option in 6.0 and brought in back
in 7.4. We happen to be in the limbo where it was unsettable
from configuration.

Bug: T309648
Change-Id: Id9cc23d46e19e31f148bbb86284a2d868372d572