Add a check that deb is unreleased in prepare_commit
We have managed to make it to the stage of building debs multiple
times while forgetting to update the BUILD_NUMBER. Add a step to
the prepare_commit workflow, which should only be used when preparing
a new deb to be built, to check if the referenced deb has already
been released to apt.wikimedia.org.

This isn't the strongest check, apt.wikimedia.org doesn't keep a full
history of releases, mostly this is checking if the version being
built is the already released version, but hopefully it's enough to
help us catch errors earlier.

Change-Id: I16f654e0958907d9e6dacf9e7fb77501522cf2d1